Fixed sentence "For every input made.."
Changed sentence "For every input made by the user, we update our board object accordingly, ie by manipulating the status of the affected cell" to "For every user input, the status of affected cell is changed and the board object is updated accordingly. to make a concise sentence and to remove pronouns.