Remove ambiguity in first 3 lines of 2nd paragraph
Changed the sentence "we have designed a class called Tank that is used to instantiate each tank on the board." to "a Tank class is designed to instantiate each Tank object on the Board object." to remove pronouns and to make it clear what "tank on the board" means.
Added a comma at the end of "To handle the active tank collection and fortress health". 
"Each tank" is changed to "each Tank object" for clarity. Added a full stop at the end of this sentence.