Fixed TankCollection description
changed sentence "We then create a class called TankCollection that holds the entire collection of tanks in it" to "a TankCollection class is designed to hold entire collection of Tank objects." to remove pronouns. 
"create" is changed to "designed" to be consistent with the previous class description.
"Each" is added before "TankCollection" and "object" is added after it to be consistent with previous object attribute description.
"Based the cumulative attributes" is incorrect.